[add]Standalone Activity Statistics page， update AgendaModel
