+++ components/Activity/Charts.module.less
@@ -1,0 +1,3 @@
+.content {
+  margin-top: 10px;
+}
