+++ components/Activity/Charts.tsx
@@ -9,6 +9,7 @@ import {
   YAxis,
 } from 'echarts-jsx';
 
+import styles from './Charts.module.less';
 import { i18n } from '../../models/Base/Translation';
 
 const { t } = i18n;
@@ -18,38 +19,41 @@ type ActivityDataProps = {
   mentorOrganizationCounts: Record<string, number>;
 };
 
-const ActivityCharts: FC<ActivityDataProps> = observer(props => {
-  const { keynoteSpeechCounts, mentorOrganizationCounts } = props;
-  const keynoteSpeechList = Object.entries(keynoteSpeechCounts).sort(
-    (a, b) => b[1] - a[1],
-  );
+const ActivityCharts: FC<ActivityDataProps> = observer(
+  ({ keynoteSpeechCounts, mentorOrganizationCounts }) => {
+    const keynoteSpeechList = Object.entries(keynoteSpeechCounts).sort(
+      (a, b) => b[1] - a[1],
+    );
 
-  const mentorOrganizationList = Object.entries(mentorOrganizationCounts).sort(
-    (a, b) => b[1] - a[1],
-  );
+    const mentorOrganizationList = Object.entries(
+      mentorOrganizationCounts,
+    ).sort((a, b) => b[1] - a[1]);
 
-  return (
-    <div style={{ marginTop: '10px', minHeight: '70vh' }}>
-      <SVGCharts>
-        <Title>{t('distribution_of_activity_topics_by_heat')}</Title>
-        <XAxis type="category" data={keynoteSpeechList.map(([key]) => key)} />
-        <YAxis type="value" />
-        <BarSeries data={keynoteSpeechList.map(([{}, value]) => value)} />
-        <Tooltip />
-      </SVGCharts>
+    return (
+      <div className={styles.content}>
+        <SVGCharts>
+          <Title>{t('distribution_of_activity_topics_by_heat')}</Title>
+          <XAxis type="category" data={keynoteSpeechList.map(([key]) => key)} />
+          <YAxis type="value" />
+          <BarSeries data={keynoteSpeechList.map(([{}, value]) => value)} />
+          <Tooltip />
+        </SVGCharts>
 
-      <SVGCharts>
-        <Title>{t('distribution_of_mentor_organizations_by_topics')}</Title>
-        <XAxis
-          type="category"
-          data={mentorOrganizationList.map(([key]) => key)}
-        />
-        <YAxis type="value" />
-        <BarSeries data={mentorOrganizationList.map(([{}, value]) => value)} />
-        <Tooltip />
-      </SVGCharts>
-    </div>
-  );
-});
+        <SVGCharts>
+          <Title>{t('distribution_of_mentor_organizations_by_topics')}</Title>
+          <XAxis
+            type="category"
+            data={mentorOrganizationList.map(([key]) => key)}
+          />
+          <YAxis type="value" />
+          <BarSeries
+            data={mentorOrganizationList.map(([{}, value]) => value)}
+          />
+          <Tooltip />
+        </SVGCharts>
+      </div>
+    );
+  },
+);
 
 export default ActivityCharts;
