+++ pages/activity/[id]/charts/index.tsx
@@ -1,19 +1,9 @@
 import { FC } from 'react';
-import { observer } from 'mobx-react';
-import {
-  BarSeries,
-  SVGCharts,
-  Title,
-  Tooltip,
-  XAxis,
-  YAxis,
-} from 'echarts-jsx';
 import { cache, compose, errorLogger, translator } from 'next-ssr-middleware';
+import dynamic from 'next/dynamic';
 
 import { i18n } from '../../../../models/Base/Translation';
 import { ActivityModel } from '../../../../models/Activity';
-import { processAgendaData } from '../../../../models/Activity/AgendaUtils';
-import dynamic from 'next/dynamic';
 
 const ActivityCharts = dynamic(
   () => import('../../../../components/Activity/Charts'),
@@ -33,9 +23,8 @@ export const getServerSideProps = compose(
     const activityStore = new ActivityModel();
     await activityStore.getOne(params!.id, true);
 
-    const agendaGroup = await activityStore.currentAgenda!.getGroup();
     const { keynoteSpeechCounts, mentorOrganizationCounts } =
-      processAgendaData(agendaGroup);
+      await activityStore.currentAgenda!.getStatistics();
 
     return {
       props: { keynoteSpeechCounts, mentorOrganizationCounts },
@@ -43,15 +32,6 @@ export const getServerSideProps = compose(
   },
 );
 
-const Charts: FC<ActivityDataProps> = observer(props => {
-  const { keynoteSpeechCounts, mentorOrganizationCounts } = props;
-
-  return (
-    <ActivityCharts
-      keynoteSpeechCounts={keynoteSpeechCounts}
-      mentorOrganizationCounts={mentorOrganizationCounts}
-    />
-  );
-});
+const Charts: FC<ActivityDataProps> = props => <ActivityCharts {...props} />;
 
 export default Charts;
